+++ FireStep/FPDController.h
@@ -28,18 +28,19 @@ protected:
     Status finalizeProbe_MTO_FPD(JsonCommand& jcmd, JsonObject& jobj, const char* key);
     Status finalizeHome();
 	Status processCalibrateCore(JsonCommand &jcmd, JsonObject& jobj, const char* key, FPDCalibrateHome &cal);
-
-public:
-    FPDController(Machine& machine);
-    XYZ3D getXYZ3D();
-    virtual const char * name();
-    virtual void onTopologyChanged();
+    virtual Status processMark(JsonCommand& jcmd, JsonObject& jobj, const char* key);
 	virtual Status processCalibrate(JsonCommand &jcmd, JsonObject& jobj, const char* keycal);
     virtual Status processHome(JsonCommand& jcmd, JsonObject& jobj, const char* key);
     virtual Status processPosition(JsonCommand &jcmd, JsonObject& jobj, const char* key);
     virtual Status processProbe(JsonCommand& jcmd, JsonObject& jobj, const char* key);
     virtual Status processDimension(JsonCommand& jcmd, JsonObject& jobj, const char* key);
     virtual	Status processMove(JsonCommand& jcmd, JsonObject& jobj, const char* key);
+
+public:
+    FPDController(Machine& machine);
+    XYZ3D getXYZ3D();
+    virtual const char * name();
+    virtual void onTopologyChanged();
 } FPDController;
 
 } // namespace firestep